Added Build Options
Sublime-Text is able to execute a compiler from it's "Tools" menu. Therefore this file points to "$folder\scripts\ecopmile.exe" and runs it for the current file. For that to work you must have opened pol folder as first in the list of opened folders.

There is an optional "Build All" variant, that will not give the file path as a parameter and will build everything within the folder. To use it press "Ctrl+Shift+B" and select "POL_eScript - Build All", since "POL_eScript" will build only the current file.
"Ctrl+B" will also run the normal build current file.

I have only added the windows options for building.